refactor: rename lambda to not reference dataset catalog
