+++ geostore/step_function_keys.py
@@ -33,7 +33,7 @@ S3_ROLE_ARN_KEY = "s3_role_arn"
 STATUS_KEY = "status"
 STEP_FUNCTION_KEY = "step_function"
 TITLE_KEY = "title"
-UPDATE_DATASET_KEY = "update_dataset_catalog"
+UPDATE_DATASET_KEY = "update_root_catalog"
 UPLOAD_STATUS_KEY = "upload_status"
 VALIDATION_KEY = "validation"
 VERSION_ID_KEY = "version_id"
